% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{getATM}
\alias{getATM}
\title{Compute ATM}
\usage{
getATM(net, syncAttractors, MAX_STEPS_TO_FIND_ATTRACTORS = 1000)
}
\arguments{
\item{net}{Boolean network loaded with loadNetwork() of BoolNet package}

\item{syncAttractors}{synchronous attractors of the network}

\item{MAX_STEPS_TO_FIND_ATTRACTORS}{number of steps after that the dynamics after the perturbation gives up}
}
\value{
List of ATM, lostFlips, and attractors in both decimal and binary format
}
\description{
Creates a structure for constructing a sort of landscape
}
\examples{

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
getATM(net, attractors)

}
