% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{getDifferentiationTreeAsDOTString}
\alias{getDifferentiationTreeAsDOTString}
\title{Generates a DOT tree-like representation of the TESs}
\usage{
getDifferentiationTreeAsDOTString(TESs)
}
\arguments{
\item{TESs}{TES structure computed with \code{\link{getTESs}}}
}
\value{
A string that describe the computed TES-based differentiation tree by using the DOT Language.
}
\description{
\code{getDifferentiationTreeAsDOTString} returns the description of the computed TES-based differentiation tree by using the DOT Language.
The returned DOT description can be passed to the \code{\link{saveDotFileDifferentiationTree}} in order to save it to a file in the local filesystem.
}
\examples{

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- getATM(net, attractors)
TESs <- getTESs(ATM) 
dotString <- getDifferentiationTreeAsDOTString

}
