% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{getTESs}
\alias{getTESs}
\title{Compute TES}
\usage{
getTESs(ATM)
}
\arguments{
\item{ATM}{ATM structured as returned from the getATM() method}
}
\value{
List of TES, and ATM structure
}
\description{
Creates a structure for constructing the TES as described in  A Dynamical Model of Genetic Networks for Cell Differentiation 
Villani M, Barbieri A, Serra R (2011) A Dynamical Model of Genetic Networks for Cell Differentiation. PLOS ONE 6(3): e17703. https://doi.org/10.1371/journal.pone.0017703
}
\examples{

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- getATM(net, attractors)
getTESs(ATM) 

}
