% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{saveDifferentiationTreeToFile}
\alias{saveDifferentiationTreeToFile}
\title{Save the differentiation tree using the DOT syntax and, if needed, produce the related image.}
\usage{
saveDifferentiationTreeToFile(TESs, filename, saveImage = TRUE)
}
\arguments{
\item{TESs}{TES structure computed with \code{\link{getTESs}}}

\item{filename}{The filename of the .gv file}

\item{saveImage}{Logical parameter indicating whether \code{saveDotFileDifferentiationTree} have to produce also the image of the tree, in .svg format.}
}
\value{
None
}
\description{
\code{saveDifferentiationTreeToFile} saves the DOT representation of the derived differentiation tree into a file.
}
\examples{

\dontrun{
net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- getATM(net, attractors)
TESs <- getTESs(ATM) 
saveDifferentiationTreeToFile(TESs, "exampleTree")
}	

}
