% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{saveDotFileDifferentiationTree}
\alias{saveDotFileDifferentiationTree}
\title{Create TES based differentiation tree}
\usage{
saveDotFileDifferentiationTree(TESs, filename, saveImage = TRUE)
}
\arguments{
\item{TESs}{TES structure as returned from the computeTES() method}

\item{filename}{desired output filename}

\item{saveImage}{if you want already the .svg image of the tree}
}
\value{
None
}
\description{
Creates representation of a differentiation tree
}
\examples{

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- computeATM(net, attractors)
TESs <- computeTESs(ATM) 
saveDotFileDifferentiationTree(TESs, "exampleTree")

}
