% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{saveDotFileDifferentiationTree}
\alias{saveDotFileDifferentiationTree}
\title{Save differentiation tree's DOT representation into a file}
\usage{
saveDotFileDifferentiationTree(DOTRep, filename, saveImage = TRUE)
}
\arguments{
\item{DOTRep}{Representation of the TES-based differentiation tree compute using \code{\link{getDifferentiationTreeAsDOTString}}}

\item{filename}{desired output filename}

\item{saveImage}{if you want already the .svg image of the tree}
}
\value{
None
}
\description{
Save differentiation tree's DOT representation into a file usign the specified name
}
\examples{

net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- getATM(net, attractors)
TESs <- getTESs(ATM) 
dotString <- getDifferentiationTreeAsDOTString(TESs)
saveDotFileDifferentiationTree(dotString, "exampleTree")

}
