% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTES.R
\name{saveDotFileDifferentiationTree}
\alias{saveDotFileDifferentiationTree}
\title{Save differentiation tree's DOT representation into a file}
\usage{
saveDotFileDifferentiationTree(DOTRep, filename, saveImage = TRUE)
}
\arguments{
\item{DOTRep}{Representation of the TES-based differentiation tree compute using \code{\link{getDifferentiationTreeAsDOTString}}}

\item{filename}{The filename of the .gv file}

\item{saveImage}{Logical parameter indicating whether \code{saveDotFileDifferentiationTree} have to produce also the image of the tree, in .svg format.}
}
\value{
None
}
\description{
\code{saveDotFileDifferentiationTree} saves the DOT representation of the derived differentiation tree into a file.
}
\examples{

\dontrun{
net <- BoolNet::generateRandomNKNetwork(10, 2)
attractors <- BoolNet::getAttractors(net) 
ATM <- getATM(net, attractors)
TESs <- getTESs(ATM) 
dotString <- getDifferentiationTreeAsDOTString(TESs)
saveDotFileDifferentiationTree(dotString, "exampleTree")
}	

}
